Update Task-1 Number guessing game using C++
This is my first task from codsoft internship in which I have created a number guessing game using C++.
It initializes the random number generator, generates a random number between 1 and 200, and then allows the user to guess the number until they get it right. The program provides feedback if the guess is too high or too low and counts the number of attempts.